# docs/arch/API_Contract_v1.0.yaml
openapi: "3.0.3"
info:
  title: Learning OS API
  description: AI-independent learning execution system — REST API contract v1.0
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/projects:
    post:
      summary: Create a new project
      operationId: createProject
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            example:
              title: "30-Day Python Challenge"
              description: "从零开始掌握 Python 基础"
              reward: "一台机械键盘"
              reward_price: 5000
              deadline: "2026-07-08"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
              example:
                id: "a1b2c3d4"
                title: "30-Day Python Challenge"
                description: "从零开始掌握 Python 基础"
                reward: "一台机械键盘"
                reward_price: 5000
                deadline: "2026-07-08"
                progress: 0
                created_at: "2026-06-08T12:00:00Z"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

    get:
      summary: List all projects
      operationId: listProjects
      tags: [Projects]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
              example:
                projects:
                  - id: "a1b2c3d4"
                    title: "30-Day Python Challenge"
                    description: "从零开始掌握 Python 基础"
                    reward: "一台机械键盘"
                    reward_price: 5000
                    deadline: "2026-07-08"
                    progress: 0.35
                    created_at: "2026-06-08T12:00:00Z"
  /api/projects/import:
    post:
      summary: Import project from Markdown DSL
      description: |
        Parse and import a Markdown DSL document that defines a complete learning project
        with stages and tasks. Returns the created project with all its stages and tasks.
      operationId: importProject
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
            example:
              markdown: |
                # 30天 Python 入门
                从零开始掌握 Python 基础，每天一小步。

                ## 第一阶段：基础语法
                掌握 Python 基本语法和数据类型。

                ### T001 theory 20 变量与数据类型
                学习 Python 的变量命名规则和基本数据类型（int, float, str, bool）。
                > 阅读《Python编程：从入门到实践》第2章
                * https://docs.python.org/3/tutorial/introduction.html

                ### T002 practice 30 编写第一个计算器
                使用 input() 接收用户输入，实现加减乘除功能。

                ### T003 output 50 学习笔记
                整理第一阶段的知识点，输出一篇结构化笔记。

                ## 第二阶段：控制流
                掌握条件判断和循环结构。

                ### T004 theory 20 if-else 条件判断
                学习布尔表达式和条件分支。

                ### T005 practice 40 猜数字游戏
                实现一个猜数字游戏，使用 while 循环。

                ### T006 output 60 代码回顾
                总结控制流学习心得，附代码示例。

      responses:
        "201":
          description: Project imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
              example:
                project_id: "a1b2c3d4"
                title: "30天 Python 入门"
                stage_count: 2
                task_count: 6
        "400":
          description: DSL parse or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DslError"
              examples:
                parse_error:
                  value:
                    error: "PARSE_ERROR"
                    line: 5
                    message: "Invalid task line: expected '### T### type xp title'"
                validation_error:
                  value:
                    error: "VALIDATION_ERROR"
                    message: "Task T007 depends on T999 which does not exist"
  /api/projects/{pid}:
    get:
      summary: Get project detail with stages
      operationId: getProject
      tags: [Projects]
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project with stages and task summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
              example:
                id: "a1b2c3d4"
                title: "30-Day Python Challenge"
                description: "从零开始掌握 Python 基础"
                reward: "一台机械键盘"
                reward_price: 5000
                deadline: "2026-07-08"
                progress: 0.35
                created_at: "2026-06-08T12:00:00Z"
                stages:
                  - id: "s1"
                    title: "第一阶段：基础语法"
                    sort_order: 1
                    progress: 0.6
                    task_count: 3
                    done_count: 2
                  - id: "s2"
                    title: "第二阶段：控制流"
                    sort_order: 2
                    progress: 0.0
                    task_count: 3
                    done_count: 0
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"

  # ──────────────────────────────────────────────
  #  TASKS
  # ──────────────────────────────────────────────
  /api/projects/{pid}/tasks:
    get:
      summary: List all tasks in a project
      operationId: listTasks
      tags: [Tasks]
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, doing, done]
      responses:
        "200":
          description: Task list grouped by stage
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskSummary"
              example:
                tasks:
                  - id: "t1"
                    stage_id: "s1"
                    title: "变量与数据类型"
                    type: "theory"
                    xp: 20
                    status: "done"
                    sort_order: 1
                  - id: "t2"
                    stage_id: "s1"
                    title: "编写第一个计算器"
                    type: "practice"
                    xp: 30
                    status: "doing"
                    sort_order: 2
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
...
  /api/projects/{pid}/reward:
    get:
      summary: Get dream reward progress for a project
      operationId: getProjectReward
      tags: [Progress]
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dream reward progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardResponse"
              example:
                project_id: "a1b2c3d4"
                reward_name: "一台机械键盘"
                reward_price: 5000
                dream_value_earned: 35
                progress_percent: 0.7
                estimated_days_remaining: 15
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"

  # ──────────────────────────────────────────────
  #  USER
  # ──────────────────────────────────────────────
...
components:
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        reward:
          type: string
          maxLength: 200
          description: Dream reward name (e.g. "机械键盘")
        reward_price:
          type: integer
          minimum: 0
          description: Dream value points needed to claim the reward
        deadline:
          type: string
          format: date
    ImportRequest:
      type: object
      required: [markdown]
      properties:
        markdown:
          type: string
          description: Full Markdown DSL document
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        reward:
          type: string
          nullable: true
        reward_price:
          type: integer
          nullable: true
        deadline:
          type: string
          format: date
          nullable: true
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 1
        created_at:
          type: string
          format: date-time
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            stages:
              type: array
              items:
                $ref: "#/components/schemas/StageSummary"
    StageSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        sort_order:
          type: integer
        progress:
          type: number
          format: float
        task_count:
          type: integer
        done_count:
          type: integer
    ImportResponse:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        title:
          type: string
        stage_count:
          type: integer
        task_count:
          type: integer
    TaskSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [theory, practice, output]
        xp:
          type: integer
        status:
          type: string
          enum: [pending, doing, done]
        sort_order:
          type: integer
...
    RewardResponse:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        reward_name:
          type: string
          nullable: true
        reward_price:
          type: integer
          nullable: true
        dream_value_earned:
          type: integer
        progress_percent:
          type: number
          format: float
        estimated_days_remaining:
          type: integer
          nullable: true
...
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
    NotFoundError:
      type: object
      properties:
        detail:
          type: string
      example:
        detail: "Project not found"
    DslError:
      type: object
      properties:
        error:
          type: string
          enum: [PARSE_ERROR, VALIDATION_ERROR]
        line:
          type: integer
          description: Line number where error occurred (PARSE_ERROR only)
        message:
          type: string
...